WIP: fixed getblocktemplate.

+++ src/util.h
@@ -349,6 +349,6 @@ std::unique_ptr<T> MakeUnique(Args&&... args)
     return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
 }
 
-double ConvertNBitsToDiff(uint32_t nBits);
+uint64_t ConvertNBitsToDiff(uint32_t nBits);
 
 #endif // BITCOIN_UTIL_H
